od_py/requirements.txt
grpcio==1.29.0
grpcio-tools==1.29.0
numpy==1.18.5
protobuf==3.12.2
six==1.15.0
